Support Github Enterprise and OAuth.
Git parameters for this plugin should be stored under omnifocus-github
instead of github.  Under this, configuration keys looked for are:

* api - specify an API endpoint other than https://api.github.com.
  This is so you can point this at your Github Enterprise endpoint.
* user, password - A username and password pair for Basic
  http authentication.
* token - An OAuth bearer token for OAuth workflows.

If both a token and a username and password are supplied, the token
is used.

Set the parameters as such:

    $ git config --global omnifocus-github.api https://ghe.my.com

Other changes:

* Include "PR" at the front of issues which are pull requests.
* Include the full issue body in the omnifocus issue.  Useful
  to have more data in OF in case your Github endpoint is on
  the wrong side of a firewall.

TODO: Support for multiple Github endpoints and/or accounts.