Multi-account support.
The git config key of omnifocus-github.accounts can
be specified as a whitespaces separated list of github
accounts.  By default, an account name of "github"
is used.

For each account, the API endpoint and authentication
parameters can be stored in the global git config
under a key matching the account.  For example, setting
up username/password authentication for gethub.com
and OAuth token authentication for a github enterprise
installation:

  git config --global github.user me
  git config --global github.password mypassword
  git config --global myghe.api https://ghe.mydomain.com/api/v3
  git config --global myghe.token 1a2b3c4d5e6f7e8d

Then tell omnifocus-github about them:

  git config --global omnifocus-github.accounts "github myghe"

NOTE: This reverts the *default* lookup of user credentials
to the github.