Query git without forcing --global
Git allows you to `include` stuff in configs (to "assemble" multiple git config files) , but it doesn't do it when you call git with `--global` (or any other restriction like that). It'd simply ignore your includes. If we just need to query git for a config value, best to do it without imposing where they should be stored.